update Course Service for final merge into the main branch
